typing brython modules
